+++ src/views/propinas/data.js
@@ -1,0 +1,79 @@
+export const propinasData = [
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+  {
+    id: '#',
+    classe: ' 10',
+    name: 'Mário Domingos Varela ',
+    lastMonthPaid: 'Julho',
+    lastPaymentDate: '10/09/1990 ',
+  },
+]
