+++ src/views/propinas/new/data.js
@@ -1,0 +1,72 @@
+const formasDePagamento = [
+  {
+    value: 'Cash',
+    label: 'Dinheiro Vivo',
+  },
+  {
+    value: 'TPA',
+    label: 'TPA',
+  },
+  {
+    value: 'Depósito Bancário',
+    label: 'Depósito Bancário',
+  },
+  {
+    value: 'Tranferência bancária',
+    label: 'Tranferência bancária',
+  },
+  {
+    value: 'Pagamento por referência MCX',
+    label: 'Pagamento por referência MCX',
+  },
+]
+
+const meses = [
+  {
+    value: 'Janeiro',
+    label: 'Janeiro',
+  },
+  {
+    value: 'Fevereiro',
+    label: 'Fevereiro',
+  },
+  {
+    value: 'Março',
+    label: 'Março',
+  },
+  {
+    value: 'Abril',
+    label: 'Abril',
+  },
+  { value: 'Maio', label: 'Maio' },
+  { value: 'Junho', label: 'Junho' },
+  { value: 'Julho', label: 'Julho' },
+  { value: 'Agosto', label: 'Agosto' },
+  { value: 'Setembro', label: 'Setembro' },
+  { value: 'Outubro', label: 'Outubro' },
+  { value: 'Novembro', label: 'Novembro' },
+  { value: 'Dezembro', label: 'Dezembro' },
+]
+
+const anos = [
+  {
+    value: 2022,
+    label: '2022',
+  },
+  {
+    value: 2023,
+    label: '2023',
+  },
+  {
+    value: 2024,
+    label: '2024',
+  },
+  {
+    value: 2025,
+    label: '2025',
+  },
+  { value: 2026, label: '2026' },
+  { value: 2027, label: '2027' },
+]
+
+export { anos, meses, formasDePagamento }
